+++ src/index.ts
@@ -222,6 +222,7 @@ export * from "./components/EditableTitle";
 export * from "./components/ActionListItem";
 export * from "./components/ActionList";
 export * from "./components/CensusView";
+export * from "./components/Auth";
 
 // Styles - import this in your app: import "solid-ui-components/styles.css"